feat: add Stream, Radix tree and XADD command (#7)
* refactor: keyval for storing more types

* feat: add radix tree

* feat: improve radix tree to use ART

* feat: improve the performance of the Range func

* feat: further optimize art

* refactor: refactor server to be single-threaded

* feat: add functions for generating IDs, trimming by size, trimming by minid

* feat: add Stream, Radix tree and XADD command

* fix: lint err

+++ pkg/command/hello.go
@@ -22,7 +22,7 @@ type Hello struct {
 func (h *Hello) Execute(c *client.Client, wr *resp.Writer, args []*resp.Resp) error {
 	protover := resp.DefaultVersion
 	if len(args) > 0 {
-		pv, err := args[0].Int()
+		pv, err := args[0].Int64()
 		if err != nil {
 			return wr.WriteError(errors.New("invalid protocol version"))
 		}
